Move the 'ResolutionContext' type into its own header
