+++ frontend/include/chpl/resolution/resolution-types.h
@@ -22,6 +22,7 @@
 
 #include "chpl/framework/query-impl.h"
 #include "chpl/framework/UniqueString.h"
+#include "chpl/resolution/ResolutionContext.h"
 #include "chpl/resolution/scope-types.h"
 #include "chpl/types/CompositeType.h"
 #include "chpl/types/EnumType.h"
@@ -40,14 +41,8 @@
 #include <utility>
 
 namespace chpl {
-
-namespace parsing {
-  bool idIsNestedFunction(Context* context, ID id);
-}
-
 namespace resolution {
 
-class ResolutionContext;
 using SubstitutionsMap = types::CompositeType::SubstitutionsMap;
 
 /**
@@ -1941,622 +1936,9 @@ class CallScopeInfo {
   const PoiScope* poiScope() const { return poiScope_; }
 };
 
-class ResolvedFunction;
-struct Resolver;
-
-/**
-  This class is used to manage mutable state that may be required while
-  resolving. Any resolution function that requires this state will take
-  a 'ResolutionContext*' as its first argument instead of a 'Context'.
-  In the vast majority of cases a user can call such a function using
-  the following sort of pattern:
-
-    ResolutionContext rcval(context);
-    auto crr = resolveCall(&rcval, ...);
-
-  The 'ResolutionContext' consists of a series of 'frames' that roughly
-  correspond to typical stack frames.
-*/
-class ResolutionContext {
- private:
-  // This class is used to implement a cache for stored results.
-  class StoredResultBase {
-   public:
-    StoredResultBase() = default;
-    virtual ~StoredResultBase() = default;
-    virtual void* get() { return nullptr; }
-    virtual bool operator==(const StoredResultBase& rhs) const = 0;
-    virtual size_t hash() const = 0;
-    struct OwnedKeyByValEquals {
-      bool operator()(const owned<StoredResultBase>& lhs,
-                      const owned<StoredResultBase>& rhs) const {
-        return *lhs == *rhs;
-      }
-    };
-    struct OwnedKeyByValHash {
-      size_t operator()(const owned<StoredResultBase>& x) const {
-        return x->hash();
-      }
-    };
-  };
-
-  // This class is used to implement a cache for stored results.
-  template <typename T>
-  class StoredResult final : public StoredResultBase {
-    T t_;
-   public:
-    StoredResult(T&& t) : t_(std::move(t)) {}
-    virtual ~StoredResult() = default;
-    virtual void* get() override { return &t_; }
-    virtual size_t hash() const override { return chpl::hash(t_); }
-    virtual bool operator==(const StoredResultBase& rhs) const override {
-      auto rptr = ((const StoredResult<T>*) &rhs);
-      return this->t_ == rptr->t_;
-    }
-  };
-
- public:
-
-  /** This class represents a single 'frame' for a symbol. It is currently
-      a thin wrapper around the internal 'Resolver' type. Generally, a
-      frame is pushed every time specific resolvers are created, and popped
-      when they are destroyed. */
-  class Frame {
-   public:
-    enum Kind { FUNCTION, MODULE, UNKNOWN };
-
-   private:
-    friend class ResolutionContext;
-    using StoreSlot = owned<StoredResultBase>;
-    using StoreHash = StoredResultBase::OwnedKeyByValHash;
-    using StoreEqual = StoredResultBase::OwnedKeyByValEquals;
-    using Store = std::unordered_set<StoreSlot, StoreHash, StoreEqual>;
-    static constexpr int64_t BASE_FRAME_INDEX = -8;
-    static const ID EMPTY_AST_ID;
-
-    Resolver* rv_ = nullptr;
-    const ResolvedFunction* rf_ = nullptr;
-    int64_t index_ = BASE_FRAME_INDEX;
-    Store cachedResults_;
-    Kind kind_ = UNKNOWN;
-
-    Frame() = default;
-    Frame(Resolver* rv, Kind kind, int64_t index)
-      : rv_(rv), index_(index), kind_(kind) {
-    }
-    Frame(const ResolvedFunction* rf, int64_t index)
-      : rf_(rf), index_(index), kind_(FUNCTION) {
-    }
-
-   public:
-   ~Frame() = default;
-    Frame(const Frame& rhs) = delete;
-    Frame(Frame&& rhs) = default;
-    Frame& operator=(Frame&& rhs) = default;
-    Frame& operator=(const Frame& rhs) = delete;
-
-    const Frame* parent(const ResolutionContext* rc) const {
-      if ((index_ - 1) < 0) return nullptr;
-      return &rc->frames_[index_ - 1];
-    }
-    Frame* parent(ResolutionContext* rc) {
-      if ((index_ - 1) < 0) return nullptr;
-      return &rc->frames_[index_ - 1];
-    }
-
-    Resolver* rv() { return rv_; }
-    const ResolvedFunction* rf() { return rf_; }
-
-    bool isEmpty() { return !rv() && !rf(); }
-    const ID& id() const;
-    const TypedFnSignature* signature() const;
-    const ResolutionResultByPostorderID* resolutionById() const;
-    bool isUnstable() const;
-
-    template <typename T>
-    const T& cache(T t) {
-      auto v = toOwned(new StoredResult<T>(std::move(t)));
-      auto it = cachedResults_.insert(std::move(v)).first;
-      StoredResultBase* base = it->get();
-      CHPL_ASSERT(base);
-      auto ptr = static_cast<T*>(base->get());
-      return *ptr;
-    }
-  };
-
- private:
-  friend struct Resolver;
-
-  Context* context_ = nullptr;
-  std::vector<Frame> frames_;
-  int numUnstableFrames_ = 0;
-  Frame baseFrame_;
-
-  const Frame* pushFrame(const ResolvedFunction* rf);
-  const Frame* pushFrame(Resolver* rv, Frame::Kind kind);
-  void popFrame(const ResolvedFunction* rf);
-  void popFrame(Resolver* rv);
-
-  Frame* lastFrameMutable() {
-    return frames_.empty() ? nullptr : &frames_.back();
-  }
-  Frame& lastFrameOrBaseMutable() {
-    return frames_.empty() ? baseFrame_ : frames_.back();
-  }
-
-  template <typename F>
-  Frame* findFrameMatchingMutable(F predicate) {
-    for (auto f = lastFrameMutable(); f; f = f->parent(this)) {
-      if (f && predicate(*f)) return f;
-    }
-    return nullptr;
-  }
-
-  Resolver* findParentResolverFor(const TypedFnSignature* sig) {
-    if (!sig || !sig->parentFn()) return nullptr;
-    auto fptr = findFrameMatchingMutable([=](auto& f) {
-      return !f.isEmpty() && f.isUnstable() &&
-              f.signature() == sig->parentFn() &&
-              f.rv() != nullptr;
-    });
-    return fptr ? fptr->rv() : nullptr;
-  }
-
- public:
-  /** Create an empty 'ResolutionContext'. */
-  ResolutionContext(Context* context) : context_(context) {}
-  ResolutionContext(const ResolutionContext& rhs) = delete;
-  ResolutionContext(ResolutionContext&& rhs) = default;
- ~ResolutionContext() = default;
-
-  /** Get the underlying compiler context. */
-  Context* context() const { return context_; }
-
-  /** If 'true', then this 'ResolutionContext' has no state available. */
-  bool isEmpty() const { return frames_.empty(); }
-
-  /** If 'true', then this contains one or more unstable frames. */
-  bool isUnstable() const { return numUnstableFrames_ != 0; }
-
-  /** Get the last frame from the frame stack. */
-  const Frame* lastFrame() const {
-    return frames_.empty() ? nullptr : &frames_.back();
-  }
-
-  /** Get the frame stack. */
-  const std::vector<Frame>& frames() const { return frames_; }
-
-  /** Get the base frame. It is not on the frame stack. */
-  const Frame& baseFrame() const { return baseFrame_; }
-
-  /** Find the first stack frame matching a predicate in LIFO order. */
-  template <typename F>
-  const Frame* findFrameMatching(F predicate) const {
-    for (auto f = lastFrame(); f; f = f->parent(this)) {
-      if (f && predicate(*f)) return f;
-    }
-    return nullptr;
-  }
-
-  /** Find the first stack frame containing _or_ matching an id. */
-  const Frame* findFrameWithId(const ID& id) const {
-    return findFrameMatching([&](auto& f) {
-      return f.id() == id || f.id().contains(id);
-    });
-  }
-
-  /** Determine if the compiler context query cache can be invoked. */
-  template <typename T>
-  bool canUseGlobalCacheConsidering(const T& t) const {
-    if (!isUnstable()) return true;
-    return ResolutionContext::canUseGlobalCache(context_, t);
-  }
-
-  /** Can the global query cache be invoked considering 't' and this? */
-  template <typename T>
-  bool canUseGlobalCacheConsidering(const T* t) const {
-    return canUseGlobalCacheConsidering(*t);
-  }
-  template <typename... Ts>
-  bool canUseGlobalCacheConsidering(const std::tuple<Ts...>& ts) {
-    return std::apply([=](auto&... x) {
-      return (canUseGlobalCacheConsidering(x) && ...);
-    }, ts);
-  }
-
-  // This overload takes at least two items to differentiate itself from
-  // the single argument overloads above (it is possible for the parameter
-  // pack to be empty).
-  template <typename T1, typename T2, typename... Ts>
-  bool canUseGlobalCacheConsidering(const T1& t1, const T2& t2,
-                                    const Ts& ...ts) {
-    if (isEmpty() || !isUnstable()) return true;
-    auto tup = std::forward_as_tuple(t1, t2, ts...);
-    return canUseGlobalCacheConsidering(tup);
-  }
-
-  /** Can the global query cache can be invoked considering only 't'?
-      By default, assume it can be. */
-  template <typename T>
-  static bool canUseGlobalCache(Context* context, const T& t) {
-    return true;
-  }
-  template <typename T>
-  static bool canUseGlobalCache(Context* context, const T*& t) {
-    if (t == nullptr) return true;
-    return canUseGlobalCache(context, *t);
-  }
-  template <typename T>
-  static bool canUseGlobalCache(Context* context, const T* const& t) {
-    if (t == nullptr) return true;
-    return canUseGlobalCache(context, *t);
-  }
-  template <typename T>
-  static bool canUseGlobalCache(Context* context, const std::vector<T>& v) {
-    for (auto& x : v) if (!canUseGlobalCache(context, x)) return false;
-    return true;
-  }
-
-  static bool
-  canUseGlobalCache(Context* context, const UntypedFnSignature& t) {
-    return canUseGlobalCache(context, t.id());
-  }
-  static bool
-  canUseGlobalCache(Context* context, const TypedFnSignature& t) {
-    return !t.isNestedFunction();
-  }
-  static bool
-  canUseGlobalCache(Context* context, const ID& t) {
-    return !parsing::idIsNestedFunction(context, t);
-  }
-  static bool
-  canUseGlobalCache(Context* context, const MatchingIdsWithName& ids) {
-    for (auto& x : ids) if (!canUseGlobalCache(context, x)) return false;
-    return true;
-  }
-
-  template <auto F, typename RetByVal, typename... ArgsByValue>
-  struct GlobalQueryWrapper;
-
-  template <auto F, typename InvokeRet, typename... InvokeArgs>
-  class GlobalQuery;
-
-  template <auto F, typename InvokeArgsTuple>
-  struct CanUseGlobalCache;
-
-  template <auto F, typename RetByVal, typename InvokeArgsTuple>
-  struct UnstableCache;
-
-  template <auto F, typename InvokeArgsTuple>
-  struct GlobalComputeSetup;
-
-  template <auto F, typename InvokeRet, typename... InvokeArgs>
-  class Query;
-
-  /** Use to construct an instance of a 'ResolutionContext::Query'.
-      These implement the 'CHPL_RESOLUTION_QUERY...' family of
-      macros defined below. */
-  template <auto F, typename... InvokeArgs>
-  auto createQueryClass(const char* name, InvokeArgs&&... args) {
-    using InvokeRet = decltype(F(this, args...));
-    using Q = ResolutionContext::Query<F, InvokeRet, InvokeArgs...>;
-    return Q(this, name, std::forward<InvokeArgs>(args)...);
-  }
-};
-
-/** This struct contains a "hidden" traditional query which stores the result
-    of invoking 'query' (the instantiated static function below) into the
-    global compiler context. Instantiation of 'GlobalQueryWrapper' is
-    precisely controlled by 'GlobalQuery' in order to avoid spurious
-    instantiations. Note that the actual query (with the user's computation
-    code) is the template value 'F'. The function 'query' exists solely to
-    work with the global query cache, which requires a specific signature.
-
-    The only time 'query' should ever be called directly is when the global
-    context is recomputing queries after bumping the revision. As such,
-    it just creates an empty 'ResolutionContext' and calls into 'F'
-    directly. In this state, it should bypass any of the initial lookup.
-*/
-template <auto F, typename RetByVal, typename... ArgsByValue>
-struct ResolutionContext::GlobalQueryWrapper {
-  static const RetByVal& query(Context* context, ArgsByValue... args) {
-    ResolutionContext rcval(context);
-    return F(&rcval, std::move(args)...);
-  }
-};
-
-// TODO: Consider moving all global queries to use this implementation.
-/** This is a class-based wrapper around global context queries. */
-template <auto F, typename InvokeRet, typename... InvokeArgs>
-class ResolutionContext::GlobalQuery {
-  static auto getStopwatchType() {
-    // These values are invalid, but that's OK, we just need their types.
-    Context* context = nullptr;
-    querydetail::QueryMapBase* base = nullptr;
-    ArgsByValueTuple* ap = nullptr;
-    return context->makeQueryTimingStopwatch(base, *ap);
-  }
-
- public:
-  using Ret = InvokeRet;
-  using RetNoRef = typename std::remove_reference<Ret>::type;
-  using RetByVal = typename std::remove_const<RetNoRef>::type;
-  template <typename T>
-  using Value = std::decay_t<T>;
-  using InvokeArgsTuple = std::tuple<InvokeArgs...>;
-  using ArgsByValueTuple = std::tuple<Value<InvokeArgs>...>;
-  using Wrapper = GlobalQueryWrapper<F, RetByVal, Value<InvokeArgs>...>;
-
-  #if CHPL_QUERY_TIMING_AND_TRACE_ENABLED
-    static constexpr bool STOPWATCH_IS_ACTIVE = true;
-  #else
-    static constexpr bool STOPWATCH_IS_ACTIVE = false;
-  #endif
-
- private:
-  // Exposing the implementation details of 'QUERY_BEGIN', 'QUERY_END'.
-  using QueryMap = querydetail::QueryMap<RetByVal, Value<InvokeArgs>...>;
-  using QueryRet = querydetail::QueryMapResult<RetByVal, Value<InvokeArgs>...>;
-  using RecomputeMarker = Context::RecomputeMarker;
-  using Stopwatch = decltype(getStopwatchType());
-
-  static constexpr bool RETURNS_CONST = std::is_const<RetNoRef>::value;
-  static constexpr bool RETURNS_REF = std::is_reference<Ret>::value;
-  static constexpr auto QUERY = Wrapper::query;
-
-  // Assert early that the user-called function is returning by 'const&'.
-  static_assert(RETURNS_CONST, "Query needs to return by 'const'!");
-  static_assert(RETURNS_REF, "Query needs to return by '&'!");
-
-  Context* context_ = nullptr;
-  const char* name_ = nullptr;
-
-  // References to the arguments as they were passed in.
-  InvokeArgsTuple args_;
-
-  // TODO: Right now this has to copy-in by value due to limitations
-  // of the 'query...' methods. It would be nice if we could use the
-  // 'InvokeArgsTuple' tuple type instead to avoid redundant copies.
-  ArgsByValueTuple ap_ = args_;
-
-  // This state is used to mimick 'QUERY_BEGIN' and 'QUERY_END'.
-  RecomputeMarker recomputeMarker_;
-  QueryMap* beginMap_ = nullptr;
-  const QueryRet* beginRet_ = nullptr;
-  Stopwatch stopwatch_;
-  bool isInput_ = false;
-
- public:
-  GlobalQuery(Context* context, const char* name, InvokeArgs&&... args)
-        : context_(context),
-          name_(name),
-          args_({std::forward<InvokeArgs>(args)...}),
-          stopwatch_(context_->makeQueryTimingStopwatch(nullptr, ap_)) {
-    CHPL_ASSERT(context_ != nullptr);
-  }
-
-  GlobalQuery(const GlobalQuery& rhs) = delete;
-  GlobalQuery(GlobalQuery&& rhs) = default;
-
-  const InvokeArgsTuple& args() const { return args_; }
-
-  const RetByVal* begin() {
-    // Fetch/emplace the map entries, which indicates the query is running.
-    beginMap_ = context_->queryBeginGetMap(QUERY, ap_, name_, isInput_);
-    beginRet_ = context_->queryBeginGetResult(beginMap_, ap_);
-
-    // Exit early if we've found a result for this revision.
-    if (context_->queryUseSaved(QUERY, beginRet_, name_)) {
-      auto ret = &context_->queryGetSaved(beginRet_);
-      beginMap_ = nullptr;
-      beginRet_ = nullptr;
-      return ret;
-    }
-
-    // Otherwise we are computing, so set the recompute marker.
-    const bool isRecomputing = false;
-    auto activeRecomputeMarker = context_->markRecomputing(isRecomputing);
-    std::swap(recomputeMarker_, activeRecomputeMarker);
-
-    // Set the stopwatch if it is compile-time enabled.
-    if constexpr (STOPWATCH_IS_ACTIVE) {
-      context_->queryBeginTrace(name_, ap_);
-      stopwatch_ = context_->makeQueryTimingStopwatch(beginMap_, ap_);
-    }
-
-    return nullptr;
-  }
-
-  const RetByVal& end(RetByVal x) {
-    // Restoring the marker also clears it.
-    recomputeMarker_.restore();
-
-    // Pass in the result and indicate the query has ended.
-    auto& ret = context_->queryEnd(QUERY, beginMap_, beginRet_, ap_,
-                                   std::move(x),
-                                   name_);
-    beginMap_ = nullptr;
-    beginRet_ = nullptr;
-
-    return ret;
-  }
-
-  void inactiveStore(RetByVal x) {
-    context_->querySetterUpdateResult(QUERY, ap_, std::move(x),
-                                      name_,
-                                      isInput_);
-  }
-
-  bool isRunning() {
-    return context_->isQueryRunning(QUERY, ap_);
-  }
-
-  bool hasCurrentResult() {
-    return context_->hasCurrentResultForQuery(QUERY, ap_);
-  }
-};
-
-template <auto F, typename InvokeArgsTuple>
-struct ResolutionContext::GlobalComputeSetup {
-  /** Called before the query is run, returns 'true' if prep was done. */
-  bool enter(ResolutionContext* rc, const InvokeArgsTuple& args) {
-    return false;
-  }
-  /** Called before the query returns, only if 'enter' returned 'true'. */
-  void leave(ResolutionContext* rc, const InvokeArgsTuple& args) {
-  }
-};
-
-template <auto F, typename InvokeArgsTuple>
-struct ResolutionContext::CanUseGlobalCache {
-  /** Consider 'args' and determine if the global query cache can be used. */
-  bool operator()(ResolutionContext* rc, const InvokeArgsTuple& args) {
-    return std::apply([&](auto... xs) {
-      return rc->canUseGlobalCacheConsidering(xs...);
-    }, args);
-  }
-};
-
-/** This class can be specialized to allow specific queries to behave
-    differently when the 'ResolutionContext' is unstable. */
-template <auto F, typename RetByVal, typename InvokeArgsTuple>
-struct ResolutionContext::UnstableCache {
-
-  /** If the 'ResolutionContext' is unstable and the global query cache
-      cannot be used, determine if a value stored in stack frames can be
-      reused for the given input. By default, no lookup occurs. */
-  const RetByVal*
-  fetchOrNull(ResolutionContext* rc, const InvokeArgsTuple& args) {
-    (void) rc;
-    (void) args;
-    return nullptr;
-  }
-
-  /** Implements the 'caching' of unstable query computations by storing
-      them into a 'ResolutionContext::Frame'. If no frame is suitable the
-      base frame may be used (which has a lifetime as long as the RC).
-      By default, use the closest frame. */
-  const RetByVal&
-  store(ResolutionContext* rc, RetByVal x, const InvokeArgsTuple& args) {
-    Frame& f = rc->lastFrameOrBaseMutable();
-    return f.cache(std::move(x));
-  }
-};
-
-/** This class represents an instance of a query created by the macro
-    'CHPL_RESOLUTION_QUERY_BEGIN...'. A resolution query is itself a
-    wrapper around a global query (cached by the 'Context*'), which is
-    represented by the 'GlobalQuery' type. */
-template <auto F, typename InvokeRet, typename... InvokeArgs>
-class ResolutionContext::Query {
-  // TODO: Refactor this into the more general notion of a "global query"
-  // (e.g., the 'GlobalQuery' defined above) that can be augmented with
-  // additional state or subclassed.
-  // NOTE: Use 'T' suffixes to prevent a GCC error message...
-  using GlobalQueryT = GlobalQuery<F, InvokeRet, InvokeArgs...>;
-  using RetByVal = typename GlobalQueryT::RetByVal;
-  using InvokeArgsTuple = typename GlobalQueryT::InvokeArgsTuple;
-  using UnstableCacheT = UnstableCache<F, RetByVal, InvokeArgsTuple>;
-  using CanUseGlobalCacheT = CanUseGlobalCache<F, InvokeArgsTuple>;
-  using GlobalComputeSetupT = GlobalComputeSetup<F, InvokeArgsTuple>;
-
-  ResolutionContext* rc_ = nullptr;
-  bool canUseGlobalCache_ = false;
-  bool didGlobalSetupOccur_ = false;
-  GlobalComputeSetupT globalSetup_;
-  GlobalQueryT global_;
-
- public:
-  Query(ResolutionContext* rc, const char* name, InvokeArgs&&... args)
-      : rc_(rc), global_(rc->context(), name, args... ) {
-    CHPL_ASSERT(rc_ && context());
-    canUseGlobalCache_ = CanUseGlobalCacheT{}(rc, global_.args());
-  }
-  Query(const Query& rhs) = delete;
-  Query(Query&& rhs) = default;
-
-  Context* context() { return rc_->context(); }
-
-  bool canUseGlobalCache() { return canUseGlobalCache_; }
-
-  const RetByVal* begin() {
-    // TODO: Use specialization to eliminate this branch?
-    if (canUseGlobalCache()) {
-      if (auto ret = global_.begin()) return ret;
-      didGlobalSetupOccur_ = globalSetup_.enter(rc_, global_.args());
-      return nullptr;
-    } else {
-      UnstableCacheT uc;
-      return uc.fetchOrNull(rc_, global_.args());
-    }
-  }
-
-  const RetByVal& end(RetByVal x) {
-    if (canUseGlobalCache()) {
-      auto& ret = global_.end(std::move(x));
-      if (didGlobalSetupOccur_) {
-        globalSetup_.leave(rc_, global_.args());
-        didGlobalSetupOccur_ = false;
-      }
-      return ret;
-    } else {
-      UnstableCacheT uc;
-      return uc.store(rc_, std::move(x), global_.args());
-    }
-  }
-
-  bool isGlobalQueryRunning() {
-    return global_.isRunning();
-  }
-
-  void inactiveStore(RetByVal x) {
-    if (canUseGlobalCache()) {
-      global_.inactiveStore(std::move(x));
-    } else {
-      UnstableCacheT uc;
-      uc.store(rc_, std::move(x), global_.args());
-    }
-  }
-};
-
-/** This macro can be used like 'QUERY_BEGIN', except it prevents the results
-    of a query from being cached in the context query cache if the computed
-    result could rely on state taken from the type 'ResolutionContext'.
-
-    Queries guarded by this macro must always return by 'const&'. When a
-    result is not stored in the context query cache, it will be stored in
-    a temporary cache maintained by the youngest 'ResolutionContext::Frame'
-    instead. The lifetime of temporarily cached values is the lifetime of
-    the associated 'ResolutionContext::Frame'.
-*/
-#define CHPL_RESOLUTION_QUERY_BEGIN(fn__, rc__, ...) \
-  auto rcquery__ = rc__->createQueryClass<fn__>(#fn__, __VA_ARGS__); \
-  if (auto ptr__ = rcquery__.begin()) return *ptr__;
-
-/** This macro can be used like 'QUERY_END'. It always returns a reference,
-    and in the case that the context cache is not used then the returned
-    value is stored in the cache of a 'ResolutionContext' frame. */
-#define CHPL_RESOLUTION_QUERY_END(ret__) (rcquery__.end(std::move(ret__)))
-
-/** Within a 'CHPL_RESOLUTION_QUERY...' use this to get the query handle. */
-#define CHPL_RESOLUTION_REF_TO_CURRENT_QUERY_HANDLE() (rcquery__)
-
-/** Use this to store a result for any 'CHPL_RESOLUTION_QUERY...' query. */
-#define CHPL_RESOLUTION_QUERY_STORE_RESULT(fn__, rc__, x__, ...) do { \
-    auto rcq__ = rc__->createQueryClass<fn__>(#fn__, __VA_ARGS__); \
-    rcq__.inactiveStore(std::move(x__)); \
-  } while (0)
-
-/** Check if the global part of a 'CHPL_RESOLUTION_QUERY...' is running. */
-#define CHPL_RESOLUTION_IS_GLOBAL_QUERY_RUNNING(fn__, rc__, ...) ([&]() { \
-    auto rcq__ = rc__->createQueryClass<fn__>(#fn__, __VA_ARGS__); \
-    return rcq__.isGlobalQueryRunning(); \
-  }())
-
 /**
   This type represents an associated action (for use within a
   ResolvedExpression).
-
  */
 class AssociatedAction {
  public:
@@ -2962,7 +2344,8 @@ class ResolvedFunction {
     }
   }
   size_t hash() const {
-    // TODO: Is it OK to skip the 'resolutionById' here?
+    // TODO: Ok to skip this here?
+    std::ignore = resolutionById_;
     size_t ret = chpl::hash(signature_, returnIntent_, poiInfo_, returnType_);
     for (auto& p : poiTraceToChild_) {
       ret = hash_combine(ret, chpl::hash(p.first));
