+++ frontend/lib/resolution/CMakeLists.txt
@@ -30,6 +30,7 @@ target_sources(ChplFrontend-obj
                intents.cpp
                maybe-const.cpp
                prims.cpp
+               ResolutionContext.cpp
                resolution-error-classes-list.cpp
                resolution-queries.cpp
                resolution-types.cpp
